wasm: Combine all tests under a single wasm-bindgen-test crate.
* Moved `url_params` tests from host running to wasm running.

  This is less efficient to run, but avoids the cost of building the
  dependencies of the all-is-cubes-wasm package (which is not in the
  main workspace) for any target other than wasm, which should cut CI
  build times by multiple minutes when uncached.

* Moved the contents of the `tests/browser` test target into the
  library. I think I was previously under the impression that this was
  not possible, but it works fine. This way, we only need to build one
  crate.
