+++ all-is-cubes-wasm/Cargo.toml
@@ -19,7 +19,7 @@ publish = false
 [lib]
 # Disable running as benchmark so that the default doesn't interfere with Criterion usage.
 bench = false
-# There are some host-runnable unit tests (but we might move those...)
+# Tests exist but require wasm-bindgen-test
 test = true
 # Not a published library, so no doctests
 doctest = false
